backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-multipart==0.0.6
asgiref==3.7.2
google-generativeai==0.3.2
transformers
huggingface_hub